requirements.txt
google-cloud-language
flask
flask-cors
wikipedia-api
python-dotenv
pymongo
bcrypt
passlib
flask-jwt-extended
DateTime
torch
torchvision
tochaudio
transformers
sentencepiece
evaluate
rouge
